lua: keep each type's fields with its bridge
FIELD_DESC has one consumer, the Lua bridge, so field.c moves out of
core/ and into game/lua/. Each game/<t>/fields.c folds into
game/lua/capi/<t>.c, which drops the extern const TYPE_DESC seam that
existed to span the two files.
