+++ src/meson.build
@@ -166,7 +166,6 @@ common_sources = [
   'trx/core/csv.c',
   'trx/core/enum_map.c',
   'trx/core/event_manager.c',
-  'trx/core/field.c',
   'trx/core/filesystem.c',
   'trx/core/hash.c',
   'trx/core/json/base.c',
@@ -278,7 +277,6 @@ common_sources = [
   'trx/game/game/state.c',
   'trx/game/game_buf.c',
   'trx/game/game_flow/common.c',
-  'trx/game/game_flow/fields.c',
   'trx/game/game_flow/inventory.c',
   'trx/game/game_flow/reader.c',
   'trx/game/game_flow/sequencer.c',
@@ -344,7 +342,6 @@ common_sources = [
   'trx/game/items/carrier.c',
   'trx/game/items/col.c',
   'trx/game/items/draw.c',
-  'trx/game/items/fields.c',
   'trx/game/items/manager.c',
   'trx/game/items/utils.c',
   'trx/game/items/walkable.c',
@@ -365,7 +362,6 @@ common_sources = [
   'trx/game/lara/control.c',
   'trx/game/lara/draw.c',
   'trx/game/lara/electric.c',
-  'trx/game/lara/fields.c',
   'trx/game/lara/flare.c',
   'trx/game/lara/hair.c',
   'trx/game/lara/interact.c',
@@ -431,6 +427,7 @@ common_sources = [
   'trx/game/lua/capi/strings.c',
   'trx/game/lua/capi/struct.c',
   'trx/game/lua/common.c',
+  'trx/game/lua/field.c',
   'trx/game/lua/registry.c',
   'trx/game/lua/sandbox.c',
   'trx/game/lua/utils.c',
@@ -549,7 +546,6 @@ common_sources = [
   'trx/game/objects/effects/splash.c',
   'trx/game/objects/effects/twinkle.c',
   'trx/game/objects/effects/water_sprite.c',
-  'trx/game/objects/fields.c',
   'trx/game/objects/general/ai_node.c',
   'trx/game/objects/general/alarm_sound.c',
   'trx/game/objects/general/animating.c',
@@ -739,7 +735,6 @@ common_sources = [
   'trx/game/replay/test_replay.c',
   'trx/game/rooms/common.c',
   'trx/game/rooms/draw.c',
-  'trx/game/rooms/fields.c',
   'trx/game/rooms/floor_data.c',
   'trx/game/rooms/geometry.c',
   'trx/game/rooms/triggers/camera.c',
