+++ src/trx/game/lua/capi/objects.c
@@ -1,8 +1,34 @@
+#include <trx/game/lua/field.h>
 #include <trx/game/lua/registry.h>
 #include <trx/game/lua/struct.h>
 #include <trx/game/lua/utils.h>
+#include <trx/game/objects/types.h>
 
-extern const TYPE_DESC TYPE_OBJECT;
+// An object is the definition every item of that type is cut from - a wolf's
+// radius, not this wolf's. Per-item state lives on the item; see trx.items.
+
+// clang-format off
+static const FIELD_DESC M_OBJECT_FIELDS[] = {
+    // what the level actually has
+    FIELD_RO(OBJECT, loaded),
+    FIELD_RO(OBJECT, intelligent),
+    FIELD_RO(OBJECT, mesh_count),
+    FIELD_RO(OBJECT, anim_count),
+
+    // the numbers that decide how it behaves
+    FIELD(OBJECT, radius),
+    FIELD(OBJECT, shadow_size),
+    FIELD(OBJECT, smartness),
+    FIELD(OBJECT, pivot_length),
+    FIELD(OBJECT, semi_transparent),
+
+    // Deliberately absent: every _func pointer, the mesh and animation indices,
+    // and the frame data. They are how the engine drives an object, and a script
+    // moving them would point it at the wrong meshes.
+};
+// clang-format on
+
+TYPE_DEFINE(OBJECT, M_OBJECT_FIELDS)
 
 // An object is addressed by its id, and an id is valid for the whole session -
 // there is no pool to recycle and nothing to go stale. An object the level did
