+++ src/trx/game/lua/field.c
@@ -1,4 +1,4 @@
-#include <trx/core/field.h>
+#include <trx/game/lua/field.h>
 
 #include <trx/debug.h>
